Desktop/rag chatbot/backend/requirements.txt
fastapi
uvicorn
gdown
pypdf
langchain-text-splitters
sentence-transformers
faiss-cpu
openai
python-dotenv
numpy
